password reset flow

+++ cmd/api/tokens.go
@@ -3,14 +3,19 @@ package main
 import (
 	"net/http"
 
+	validation "github.com/go-ozzo/ozzo-validation"
+	"github.com/go-ozzo/ozzo-validation/is"
 	"github.com/micahco/api/internal/data"
 )
 
-const verificationMsg = "A verification email has been sent. Please check your inbox."
+const (
+	verificationMsg  = "A verification email has been sent. Please check your inbox."
+	passwordResetMsg = "If that email address is in our database, a token to reset your password will be sent to that address."
+)
 
 // Create a verification token with registration scope and
 // mail it to the provided email address.
-func (app *application) tokensVerificaitonPost(w http.ResponseWriter, r *http.Request) error {
+func (app *application) tokensVerificaitonRegistrationPost(w http.ResponseWriter, r *http.Request) error {
 	var input struct {
 		Email string `json:"email"`
 	}
@@ -20,10 +25,16 @@ func (app *application) tokensVerificaitonPost(w http.ResponseWriter, r *http.Re
 		return err
 	}
 
+	err = validation.ValidateStruct(&input,
+		validation.Field(&input.Email, validation.Required, is.Email),
+	)
+	if err != nil {
+		return err
+	}
+
 	// This will be the consistent message. Even if a user
 	// already exists with this email, send this message.
 	msg := envelope{"message": verificationMsg}
-	scope := data.ScopeRegistration
 
 	// Check if user with email already exists
 	exists, err := app.models.User.ExistsWithEmail(input.Email)
@@ -37,7 +48,7 @@ func (app *application) tokensVerificaitonPost(w http.ResponseWriter, r *http.Re
 	}
 
 	// Check if a verification token has already been created recently
-	exists, err = app.models.VerificationToken.Exists(scope, input.Email, nil)
+	exists, err = app.models.VerificationToken.Exists(data.ScopeRegistration, input.Email, nil)
 	if err != nil {
 		return err
 	}
@@ -47,7 +58,7 @@ func (app *application) tokensVerificaitonPost(w http.ResponseWriter, r *http.Re
 		return app.writeJSON(w, http.StatusOK, msg, nil)
 	}
 
-	t, err := app.models.VerificationToken.New(scope, input.Email, nil)
+	t, err := app.models.VerificationToken.New(data.ScopeRegistration, input.Email, nil)
 	if err != nil {
 		return err
 	}
@@ -58,18 +69,13 @@ func (app *application) tokensVerificaitonPost(w http.ResponseWriter, r *http.Re
 			"token": t.Plaintext,
 		}
 
-		err = app.mailer.Send(input.Email, "registration.tmpl", data)
-		if err != nil {
-			return err
-		}
-
-		return nil
+		return app.sendMail(input.Email, "registration.tmpl", data)
 	})
 
 	return app.writeJSON(w, http.StatusOK, msg, nil)
 }
 
-func (app *application) tokensVerificaitonUserEmailPost(w http.ResponseWriter, r *http.Request) error {
+func (app *application) tokensVerificaitonEmailChangePost(w http.ResponseWriter, r *http.Request) error {
 	var input struct {
 		Email string `json:"email"`
 	}
@@ -79,10 +85,16 @@ func (app *application) tokensVerificaitonUserEmailPost(w http.ResponseWriter, r
 		return err
 	}
 
+	err = validation.ValidateStruct(&input,
+		validation.Field(&input.Email, validation.Required, is.Email),
+	)
+	if err != nil {
+		return err
+	}
+
 	// This will be the consistent message. Even if a user
 	// already exists with this email, send this message.
 	msg := envelope{"message": verificationMsg}
-	scope := data.ScopeChangeEmail
 
 	// Check if user with email already exists
 	exists, err := app.models.User.ExistsWithEmail(input.Email)
@@ -99,7 +111,7 @@ func (app *application) tokensVerificaitonUserEmailPost(w http.ResponseWriter, r
 	user := app.contextGetUser(r)
 
 	// Check if a verification token has already been created recently
-	exists, err = app.models.VerificationToken.Exists(scope, input.Email, &user.ID)
+	exists, err = app.models.VerificationToken.Exists(data.ScopeEmailChange, input.Email, &user.ID)
 	if err != nil {
 		return err
 	}
@@ -110,7 +122,7 @@ func (app *application) tokensVerificaitonUserEmailPost(w http.ResponseWriter, r
 	}
 
 	// Create verification token for user with new email address
-	t, err := app.models.VerificationToken.New(scope, input.Email, &user.ID)
+	t, err := app.models.VerificationToken.New(data.ScopeEmailChange, input.Email, &user.ID)
 	if err != nil {
 		return err
 	}
@@ -121,12 +133,66 @@ func (app *application) tokensVerificaitonUserEmailPost(w http.ResponseWriter, r
 			"token": t.Plaintext,
 		}
 
-		err = app.mailer.Send(input.Email, "change-email.tmpl", data)
-		if err != nil {
-			return err
+		return app.sendMail(input.Email, "email-change.tmpl", data)
+	})
+
+	return app.writeJSON(w, http.StatusOK, msg, nil)
+}
+
+func (app *application) tokensVerificaitonPasswordResetPost(w http.ResponseWriter, r *http.Request) error {
+	var input struct {
+		Email string `json:"email"`
+	}
+
+	err := app.readJSON(r, &input)
+	if err != nil {
+		return err
+	}
+
+	err = validation.ValidateStruct(&input,
+		validation.Field(&input.Email, validation.Required, is.Email),
+	)
+	if err != nil {
+		return err
+	}
+
+	// This will be the consistent message. Even if a user
+	// already exists with this email, send this message.
+	msg := envelope{"message": verificationMsg}
+
+	// Check if user with email exists
+	exists, err := app.models.User.ExistsWithEmail(input.Email)
+	if err != nil {
+		return err
+	}
+	if !exists {
+		// User with email does not exist. Still send the same
+		// message.
+		return app.writeJSON(w, http.StatusOK, msg, nil)
+	}
+
+	// Check if a verification token has already been created recently
+	exists, err = app.models.VerificationToken.Exists(data.ScopePasswordReset, input.Email, nil)
+	if err != nil {
+		return err
+	}
+	if exists {
+		// Recent verification sent, don't mail another
+		return app.writeJSON(w, http.StatusOK, msg, nil)
+	}
+
+	t, err := app.models.VerificationToken.New(data.ScopePasswordReset, input.Email, nil)
+	if err != nil {
+		return err
+	}
+
+	// Mail the plaintext token to the user's email address
+	app.background(func() error {
+		data := map[string]any{
+			"token": t.Plaintext,
 		}
 
-		return nil
+		return app.sendMail(input.Email, "password-reset.tmpl", data)
 	})
 
 	return app.writeJSON(w, http.StatusOK, msg, nil)
@@ -139,6 +205,14 @@ func (app *application) tokensAuthenticationPost(w http.ResponseWriter, r *http.
 	}
 
 	err := app.readJSON(r, &input)
+	if err != nil {
+		return err
+	}
+
+	err = validation.ValidateStruct(&input,
+		validation.Field(&input.Email, validation.Required, is.Email),
+		validation.Field(&input.Password, validation.Required, data.PasswordLength),
+	)
 	if err != nil {
 		return err
 	}
